Employee controller validation

+++ src/controllers/v1/employee.controller.ts
@@ -1,13 +1,13 @@
 import express, { Request, RequestHandler, Response, Router } from 'express';
+import validator from 'validator';
 import Employee from '../../database/models/employee.model';
 
 // GET /employees
 const getEmployees: RequestHandler = async (req: Request, res: Response) => {
   const { email } = req.query;
-  // TODO: Validate email
   if (email) {
+    if (!validator.isEmail(email as string)) return res.sendStatus(400);
     const employee = await Employee.findOne({ where: { email: email as string } });
-    // TODO: Error response
     if (!employee) return res.sendStatus(404);
     return res.json(employee);
   } else {
@@ -19,9 +19,8 @@ const getEmployees: RequestHandler = async (req: Request, res: Response) => {
 // GET /employees/:id
 const getEmployeeById: RequestHandler = async (req: Request, res: Response) => {
   const { id } = req.params;
-  // TODO: Validate id
+  if (!validator.isUUID(id, 4)) return res.sendStatus(400);
   const employee = await Employee.findOne({ where: { id } });
-  // TODO: Error response
   if (!employee) return res.sendStatus(404);
   return res.json(employee);
 };
@@ -29,7 +28,12 @@ const getEmployeeById: RequestHandler = async (req: Request, res: Response) => {
 // POST /employees
 const createEmployee: RequestHandler = async (req: Request, res: Response) => {
   const { name, email } = req.body;
-  // TODO: Validate name and email
+  if (
+    validator.isEmpty(name, { ignore_whitespace: true }) ||
+    !validator.isLength(name, { min: 3, max: 255 }) ||
+    !validator.isEmail(email)
+  )
+    return res.sendStatus(400);
   const employee = await Employee.create({ name, email });
   res.status(201).json(employee);
 };
@@ -38,9 +42,15 @@ const createEmployee: RequestHandler = async (req: Request, res: Response) => {
 const updateEmployee: RequestHandler = async (req: Request, res: Response) => {
   const { id } = req.params;
   const { name, email } = req.body;
-  // TODO: Validate id, name and email
+  if (!validator.isUUID(id, 4)) return res.sendStatus(400);
+  if (
+    !validator.isUUID(id, 4) ||
+    validator.isEmpty(name, { ignore_whitespace: true }) ||
+    !validator.isLength(name, { min: 3, max: 255 }) ||
+    !validator.isEmail(email)
+  )
+    return res.sendStatus(400);
   const employee = await Employee.findOne({ where: { id } });
-  // TODO: Error response
   if (!employee) return res.sendStatus(404);
   await Employee.update({ name, email }, { where: { id } });
   return res.sendStatus(202);
@@ -49,9 +59,8 @@ const updateEmployee: RequestHandler = async (req: Request, res: Response) => {
 // DELETE /employees/:id
 const deleteEmployee: RequestHandler = async (req: Request, res: Response) => {
   const { id } = req.params;
-  // TODO: Validate id
+  if (!validator.isUUID(id, 4)) return res.sendStatus(400);
   const employee = await Employee.findOne({ where: { id } });
-  // TODO: Error response
   if (!employee) return res.sendStatus(404);
   await Employee.destroy({ where: { id } });
   return res.sendStatus(204);
